Made ModContainer an interface, added check for duplicate mod IDs
